ExceptionRepresentationParser wrapper as Angular.JS provider
This part of software is something I enjoy to be the developer of.
It is the frontend part which parses the exception error codes
sent by the very specific backend of glpi-plugin-directory.
It parses values of types :
 + int
 + string
 + single random keyword
And keys as
 + single random keyword

the parseExceptionRepresentation method returns an object like
{
  "name": "ERROR_CODE",
  "args": {
    "one": "value of first arg",
    "two": "valueofsecondarg"
    "three": 3
  }
}

for an exception like

ERROR_CODE(first="value of first arg", two=valueofsecondarg, three=3)